sdk: reinstate the message package for message accessors.
